requirements.txt
flask
python-dotenv  # for loading .env if you have secrets
gunicorn       # production WSGI server
